reworking perf debug class
